Continue to bubble logic up to top level
