+++ scripts/postinstall.ts
@@ -1,11 +1,13 @@
 import { execFileSync } from 'child_process';
 import os from 'os';
+import path from 'path';
+import fs from 'fs';
 
-import DependencyVersions from 'scripts/download/dependencies';
 import downloadLima from 'scripts/download/lima';
 import downloadMobyOpenAPISpec from 'scripts/download/moby-openapi';
 import downloadDependencies from 'scripts/download/tools';
-import downloadWSL from 'scripts/download/wsl';
+import { downloadWSLDistro, downloadHostResolverHost, downloadHostResolverPeer } from 'scripts/download/wsl';
+import { DependencyPlatform, DependencyVersions, DownloadContext, Platform, KubePlatform } from 'scripts/lib/dependencies';
 
 async function runScripts(): Promise<void> {
   // load desired versions of dependencies
@@ -15,19 +17,55 @@ async function runScripts(): Promise<void> {
   await downloadMobyOpenAPISpec();
   switch (os.platform()) {
   case 'linux':
-    await downloadDependencies('linux', depVersions);
+    const linuxDownloadContext = buildDownloadContextFor('linux');
+    await downloadDependencies(linuxDownloadContext, depVersions);
     await downloadLima();
     break;
   case 'darwin':
-    await downloadDependencies('darwin', depVersions);
+    const macosDownloadContext = buildDownloadContextFor('darwin');
+    await downloadDependencies(macosDownloadContext, depVersions);
     await downloadLima();
     break;
   case 'win32':
-    await downloadDependencies('win32', depVersions);
-    await downloadDependencies('wsl', depVersions);
-    await downloadWSL();
+    // download things for windows
+    const windowsDownloadContext = buildDownloadContextFor('win32');
+    await downloadDependencies(windowsDownloadContext, depVersions);
+    await downloadWSLDistro(windowsDownloadContext, depVersions.WSLDistro);
+    await downloadHostResolverHost(windowsDownloadContext, depVersions.hostResolver);
+
+    // download things that go inside WSL distro
+    const wslDownloadContext = buildDownloadContextFor('wsl');
+    await downloadDependencies(wslDownloadContext, depVersions);
+    await downloadHostResolverPeer(wslDownloadContext, depVersions.hostResolver);
     break;
   }
+}
+
+function buildDownloadContextFor(rawPlatform: DependencyPlatform): DownloadContext {
+  const platform = rawPlatform === 'wsl' ? 'linux' : rawPlatform;
+  const resourcesDir = path.join(process.cwd(), 'resources', platform);
+  const downloadContext: DownloadContext = {
+    dependencyPlaform: rawPlatform,
+    platform,
+    kubePlatform:      getKubePlatform(platform),
+    resourcesDir:      resourcesDir,
+    binDir:            path.join(resourcesDir, 'bin'),
+    internalDir:       path.join(resourcesDir, 'internal'),
+  };
+
+  fs.mkdirSync(downloadContext.binDir, { recursive: true });
+  fs.mkdirSync(downloadContext.internalDir, { recursive: true });
+
+  return downloadContext;
+}
+
+function getKubePlatform(platform: Platform): KubePlatform {
+  const platformToKubePlatfom: Record<Platform, KubePlatform> = {
+    'darwin': 'darwin',
+    'linux': 'linux',
+    'win32': 'windows',
+  }
+  return platformToKubePlatfom[platform];
 }
 
 runScripts().then(() => {
